Added a T_EntityId typedef, and added it as parameter for Entity.
Changed Entity so that Id, Type and Name now are properties. getId, getType and getName are still present as convenient methods on the Entity class.

Took out some deprecated code from Component. It was easy to use this wrong and end up creating unique properties for each component... it covered a corner-case that's solved using PropertyList now.