A small change in the process of being built concerning events. The old event system will remain, called TemplateEventHandler, but I'll add an additional one called AnyEventHandler.
This event handler treats all arguments as Any, a typesafe construct seen in libraries like Boost, and expect the event receiver of casting the arguments to correct type. The cool thing with this event handler, is that it will support scheduled events with a non-intrusive implementation! 

I didn't implement a scheduler on the Template-based Event Handler, because I couldn't find a non-intrusive approach. Will probably have to integrate a enumerate list or some such :)