Fixed a small bug in the delegates example which entered const type & into the template fields in some places. Since the delegate handler implementation is built so that it does const T & internally, this caused havok on VS2008. On VS2010 it was no problem at all... it was still a bug though, so good to crush it out of the svn head.
Further, vs2008 wouldn't accept std::hash_value. stdext::hash_value was ok though, so replaced it with that. Will have to test in vs2010 to see if this is ok, or if I'll have to start ifdef'ing msvc versions in code (ugly).

Last, I added a resize feature to the PropertyList, expanding on the basic std::vector features a bit.