Just added some better error message when you try to create a component type that's not found in the registered list of component creator function pointers.
This can occur if the input type is spelled wrong, or when the number of optional/custom parameters differ from the register to the create function.

factory.registerComponentCustom1<int>("SomeComponent");

entity.addComponent("SomeComponent") //<-- misses the custom parameter, and should throw an error.

Also, there's the case of trying to create a component who's custom parameter(s) differ in type from those registered.