Added propertyAdded signals with and without UserData to PropertyHandler. This allows external logic to access this functionality, like for instance an Editor which needs to get access to the properties added by a component.
UserData can be specified if, for instance an editor, needs some special data collected for properties, like the name of the component it was added to, or an instance to it even.

As UserData is a template set on the PropertyHandler, it's up to the end-user's implementation of the Entity class to determine the type of the UserData template used for this.