Added a method for working on an index of the data in a Property and still catching the data's valueChanged event callback using the same approach, basically, as PropertyList.h
The reason for this feature, was found when trying to work with matrices as data in properties, where the problem comes when you need to change something inside that matrix without replacing all it's data with a new matrix. Same could be said for indexing into a string, or any other "container" of data that you don't normally work on as a list (so it's a property, not a propertyList).